size unit in float

+++ RDN/RDN/Banner/RPSBannerAdapter.h
@@ -14,8 +14,8 @@ NS_ASSUME_NONNULL_BEGIN
 @interface RPSBanner : NSObject<RPSAdInfo>
 
 @property(nonatomic, readonly) NSString* html;
-@property(nonatomic, readonly) int width;
-@property(nonatomic, readonly) int height;
+@property(nonatomic, readonly) float width;
+@property(nonatomic, readonly) float height;
 
 @property(nonatomic, readonly) RPSURLString* measuredURL;
 @property(nonatomic, readonly) RPSURLString* inviewURL;
